requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-filter==24.3
django-js-asset==3.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
numpy==2.2.1
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
